Move functions into CDSW Module
